requirements.txt
alembic==1.18.4
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.11.0
asyncpg==0.31.0
asyncpg-stubs==0.31.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
email-validator==2.3.0
fastapi==0.120.4
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-anthropic==1.3.3
langchain-core==1.2.11
langchain-openai==1.1.9
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.5
langsmith==0.7.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
openai==2.20.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic-settings==2.12.0
pydantic_core==2.41.4
Pygments==2.19.2
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.2
sentry-sdk==2.43.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlmodel==0.0.33
starlette==0.49.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uuid_utils==0.14.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
